Conditional build example with features
