+++ cip25-wasm-example/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "cardano-multiplatform-lib-cip25-wasm"
+name = "cip25-wasm-example"
 version = "0.1.0"
 edition = "2018"
 
@@ -7,7 +7,7 @@ edition = "2018"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-core = { path = "../cip25", package = "cardano-multiplatform-lib-cip25" }
+cip25 = { path = "../cip25-wasm", package = "cardano-multiplatform-lib-cip25-wasm", features = ["wasm"] }
 cardano-multiplatform-lib-core-wasm = { path = "../core-wasm" }
 
 cbor_event = "2.2.0"
