+++ cip25-wasm-example/src/lib.rs
@@ -1,0 +1,1 @@
+pub use cip25::String64s;
